various fixes

+++ all/vim/cat-arcadia-files.sh
@@ -1,34 +1,43 @@
 #!/bin/zsh -e
 
-ARCADIA_ROOT_FILES_LIST_CACHED_FILE=${HOME}/.cache/arcadia_files_list.txt
+if [[ "${@: -1}" = "${ARCADIA_ROOT}" ]]; then
 
-unset get_arcadia_files_of_interest
+    ARCADIA_ROOT_FILES_LIST_CACHED_FILE=${HOME}/.cache/arcadia_files_list.txt
 
-get_arcadia_files_of_interest() {
-    local directories=("catboost" "contrib" "cv" "extsearch" "ml" "mapreduce" "library" "kernel" "quality" "sandbox" "search" "statbox" "strm" "util" "web" "yql" "ysite" "yweb" "junk/splinter");
+    unset get_arcadia_files_of_interest
 
-    for directory in ${directories}; do
-        fd -t f --color=never . ${ARCADIA_ROOT}/${directory}
-    done
+    get_arcadia_files_of_interest() {
+        local directories=("contrib/libs/intel" "contrib/libs/tf" "contrib/libs/tf-1.14" "contrib/libs/nvidia" "contrib/libs/eigen" "cv" "extsearch" "ml" "mapreduce" "library" "kernel" "quality/neural_net" "statbox/nile" "statbox/bindings" "statbox/qb2" "statbox/qb2_core" "util" "yql/library" "junk/splinter");
 
-    for directory in ${directories}; do
-        fd -t f --color=never . ${ARCADIA_ROOT_GENERATED}/${directory}
-    done
-}
+        for directory in ${directories}; do
+            if [[ -d "${ARCADIA_ROOT}/${directory}" ]]; then
+                fd -t f --color=never . ${ARCADIA_ROOT}/${directory}
+            fi
+        done
+
+        for directory in ${directories}; do
+            if [[ -d "${ARCADIA_ROOT_GENERATED}/${directory}" ]]; then
+                fd -t f --color=never . ${ARCADIA_ROOT_GENERATED}/${directory}
+            fi
+        done
+    }
 
 
-if [[ ! -f ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE} ]]; then
-    get_arcadia_files_of_interest > ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE}
-else
-    local current_time=`date +%s`;
-    local modification_time=`date -r ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE} +%s`
-    local time_diff=$current_time-$modification_time
-    if [[ $time_diff -gt 86400 ]]; then
-        rm -f ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE};
+    if [[ ! -f ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE} ]]; then
         get_arcadia_files_of_interest > ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE}
+    else
+        local current_time=`date +%s`;
+        local modification_time=`date -r ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE} +%s`
+        local time_diff=$current_time-$modification_time
+        if [[ $time_diff -gt 86400 ]]; then
+            rm -f ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE};
+            get_arcadia_files_of_interest > ${ARCADIA_ROOT_FILES_LIST_CACHED_FILE}
+        fi
     fi
+
+    # ag --ignore .git -g . $HOME/experiments
+
+    cat $ARCADIA_ROOT_FILES_LIST_CACHED_FILE
+else
+    fd -t f --color=never . ${@: -1}
 fi
-
-ag --ignore .git -g . $HOME/experiments
-
-cat $ARCADIA_ROOT_FILES_LIST_CACHED_FILE
